backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9
qrcode[pil]>=7.4
Pillow>=10.0
sendgrid>=6.10
python-decouple>=3.8
django-cors-headers>=4.3
pytz>=2024.1
